requirements.txt
decorator==3.4.0
Django==1.7.6
docutils==0.12
process-wrapper==0.1.2
psutil==2.2.1
robotframework==2.8.7
robotframework-selenium2library==1.6.0
selenium==2.45.0
